requirements.txt
biopython
git+https://github.com/MauriceKarrenbrock/PythonAuxiliaryFunctions.git #PythonAuxiliaryFunctions
git+https://github.com/prody/ProDy.git #prody
MDAnalysis

